+++ src/grafanarmadillo/find.py
@@ -13,6 +13,11 @@ from grafanarmadillo.types import (
 )
 
 
+def _query_message(query_type: str, query: str) -> str:
+	"""Format a message detailing the query."""
+	return f"type={query_type}, query={query}"
+
+
 class Finder:
 	"""Collection of methods for finding Grafana dashboards and folders."""
 
@@ -53,12 +58,11 @@ class Finder:
 		else:
 			search_result = self.api.search.search_dashboards(query=name, type_="dash-folder")
 			return exactly_one(
-				list(
-					filter(
-						lambda x: x["title"] == name,
-						map(lambda sr: self.api.folder.get_folder(sr["uid"]), search_result),
-					)
-				)
+				list(filter(
+					lambda x: x["title"] == name,
+					map(lambda sr: self.api.folder.get_folder(sr["uid"]), search_result),
+				)),
+				_query_message("folder", name),
 			)
 
 	def get_dashboard(self, folder_name, dashboard_name) -> DashboardSearchResult:
@@ -69,16 +73,22 @@ class Finder:
 		"""
 		folder_object = self.get_folder(folder_name)
 		dashboards = self._enumerate_dashboards_in_folders([str(folder_object["id"])])
-		return exactly_one(list(filter(lambda d: d["title"] == dashboard_name, dashboards)))
+		return exactly_one(
+			list(filter(lambda d: d["title"] == dashboard_name, dashboards)),
+			_query_message("dashboard", f"/{folder_name}/{dashboard_name}"),
+		)
 
 	def get_alert(self, folder_name, alert_name) -> AlertSearchResult:
 		"""Get an alert by its parent folder and alert name."""
 		folder_uid = self.get_folder(folder_name)["uid"]
 
-		return exactly_one(list(filter(
-			lambda a: a["title"] == alert_name and a["folderUID"] == folder_uid,
-			self.api.alertingprovisioning.get_alertrules_all()
-		)))
+		return exactly_one(
+			list(filter(
+				lambda a: a["title"] == alert_name and a["folderUID"] == folder_uid,
+				self.api.alertingprovisioning.get_alertrules_all()
+			)),
+			_query_message("alert", f"/{folder_name}/{alert_name}")
+		)
 
 	@staticmethod
 	def _resolve_path(path) -> Tuple[str, str]:
@@ -104,6 +114,11 @@ class Finder:
 		folder, dashboard = self._resolve_path(path)
 		return self.get_dashboard(folder, dashboard)
 
+	def get_alert_from_path(self, path) -> AlertSearchResult:
+		"""Get an alert from a string path like `/folder0/alert0`."""
+		folder, alert = self._resolve_path(path)
+		return self.get_alert(folder, alert)
+
 	def create_or_get_dashboard(
 		self, path: str
 	) -> Tuple[DashboardSearchResult, Optional[FolderSearchResult]]:
@@ -122,12 +137,39 @@ class Finder:
 		try:
 			dashboard = self.get_dashboard(folder_name, dashboard_name)
 		except ValueError:
-			dashboard = self.api.dashboard.update_dashboard(
+			self.api.dashboard.update_dashboard(
 				{
 					"dashboard": {"title": dashboard_name},
 					"folderId": folder["id"],
 					"folderUid": folder["uid"],
 				}
 			)
+			dashboard = self.get_dashboard(folder_name, dashboard_name)
 
 		return dashboard, folder
+
+	def create_or_get_alert(self, path: str) -> Tuple[AlertSearchResult, FolderSearchResult]:
+		"""
+		Get the information about an alert, or fake the metadata if it does not.
+
+		Creating an "empty" alert in Grafana requires filling in much more data than an empty dashboard,
+		including at least 1 rule.
+		We can fake that with a `math` rule that always returns 0,
+		but that's a lot of garbage data to inject.
+		"""
+		folder_name, alert_name = self._resolve_path(path)
+
+		try:
+			folder = self.get_folder(folder_name)
+		except ValueError:
+			folder = self.api.folder.create_folder(folder_name)
+
+		try:
+			alert = self.get_alert(folder_name, alert_name)
+		except ValueError:
+			alert = {
+				"folderUID": folder["uid"],
+				"title": alert_name
+			}
+
+		return alert, folder